backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
pdfplumber==0.10.3
requests==2.31.0
drf-yasg==1.21.7
dj-database-url==2.1.0
django-cors-headers==4.3.1
reportlab==4.0.0
Pillow==10.0.0